% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_init.R
\docType{data}
\name{AREA}
\alias{AREA}
\title{Vector of Functional Areas}
\format{
An object of class \code{character} of length 4.
}
\usage{
AREA
}
\description{
Common functional areas for demonstration purposes.
}
\keyword{datasets}
