% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\docType{data}
\name{BSGP}
\alias{BSGP}
\title{Vector of BSGP Stages}
\format{
An object of class \code{character} of length 4.
}
\usage{
BSGP
}
\description{
The four data stages commonly used in data processing.
}
\keyword{datasets}
