% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\docType{data}
\name{.AREA}
\alias{.AREA}
\title{[](https://coolors.co/3e074a-0f431c-fde8e9-f0f600)}
\format{
An object of class \code{character} of length 4.
}
\usage{
.AREA
}
\description{
[](https://coolors.co/3e074a-0f431c-fde8e9-f0f600)
}
\keyword{datasets}
