% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.cn_config_file_path_get}
\alias{.cn_config_file_path_get}
\title{Get config file path}
\usage{
.cn_config_file_path_get()
}
\value{
A character string representing the full path to the configuration YAML file.
}
\description{
Constructs the full path to the configuration YAML file based on the root directory
determined by the `.cn_root_dir_get` function, the folder name, and the file name constants.
The function uses the `fs` package to ensure robust and cross-platform compatibility.
}
\keyword{internal}
