% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.cn_constants_initialize}
\alias{.cn_constants_initialize}
\title{initialize global constants}
\usage{
.cn_constants_initialize()
}
\value{
NULL. This function creates constants dynamically in the global environment.
}
\description{
Creates global constants for colors, functional areas, stages, and environments by calling internal functions.
}
\keyword{internal}
