% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.cn_root_dir_get}
\alias{.cn_root_dir_get}
\title{Get root directory}
\usage{
.cn_root_dir_get()
}
\value{
A character string representing the root directory path:
- `"./inst/extdata"` if the code is being executed in a testing environment, during `devtools::check()`, or if the directory `inst/extdata` exists.
- `"."` (the current working directory) if the code is being executed in a normal project context.
}
\description{
Determines the root directory of the project or package based on the execution context.
}
\details{
This function dynamically identifies the root directory to be used in various execution contexts.
If the code is executed in a testing environment (`testthat`), during a package check
(`devtools::check()`), or if the directory `inst/extdata` exists, it will return the
`inst/extdata` path. Otherwise, it defaults to the current working directory.
}
\keyword{internal}
