% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_int.R
\name{.create_default_config_yaml}
\alias{.create_default_config_yaml}
\title{Create Default Config YAML File (Internal)}
\usage{
.create_default_config_yaml(
  config_dir = "./inst/extdata",
  key_value_list = NULL
)
}
\arguments{
\item{config_dir}{Character. The directory where the `config.yaml` file will be created. Defaults to `"./inst/extdata"`.}

\item{key_value_list}{List. A named list of key-value pairs to set in the configuration file.}
}
\description{
Internal helper function to create a default `config.yaml` file in the specified directory.
}
\examples{
.create_default_config_yaml(
  config_dir = "./inst/extdata",
  key_value_list = list(
    "project.name"        = "Pythia's Advice",
    "project.department"  = "EAISI",
    "csv_file_spec.delim" = ";"
    )
  )
}
\keyword{internal}
