% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion_int.R
\name{.di_fields_out_get}
\alias{.di_fields_out_get}
\title{Generate "Fields Out" Using a Generic Template}
\usage{
.di_fields_out_get(transformations)
}
\arguments{
\item{transformations}{A \code{data.table} (or similar) with columns
\code{FLDNM_OUT} and \code{TRNSFRM} at least.}
}
\value{
A single character string with the formatted \code{FLDNM_OUT} fields.
}
\description{
Internal wrapper function around \code{.di_fields_generic_get} specifically for
fields that reference \code{FLDNM_OUT} and \code{TRNSFRM}.
}
\keyword{internal}
