% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-handling_int.R
\name{.fh_data_dir_abs}
\alias{.fh_data_dir_abs}
\title{get data directory}
\usage{
.fh_data_dir_abs()
}
\value{
Character. The absolute path to the working directory.
}
\description{
Determines the working directory based on the project directory and the configuration file.
resolves placeholders such as `"OneDriveConsumer"` or `"OneDriveBusiness"`
back to real file paths, then appends the currently set environment to the path.
}
\keyword{internal}
