% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-handling_int.R
\name{.fh_files_metadata_list_get}
\alias{.fh_files_metadata_list_get}
\title{List Files in Directory with Metadata (Internal)}
\usage{
.fh_files_metadata_list_get(dir)
}
\arguments{
\item{dir}{Character. The directory to scan for files.}
}
\value{
A data.table with the following columns:
\describe{
  \item{FFN}{Full file name (absolute path).}
  \item{BFN}{Base file name (without extension).}
  \item{EXT}{File extension.}
  \item{MOD}{Modification time.}
}
}
\description{
Internal function to create a data.table containing information about all files
in a given directory.
}
\keyword{internal}
