% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-handling_int.R
\name{.fh_files_sync_list_get}
\alias{.fh_files_sync_list_get}
\title{Get New and Modified Files Between Two Directories (Internal)}
\usage{
.fh_files_sync_list_get(dir_src, dir_tgt)
}
\arguments{
\item{dir_src}{Character. Path to the source directory.}

\item{dir_tgt}{Character. Path to the target directory.}
}
\value{
A character vector of full file names (absolute paths) for files that are
new or modified in the source directory.
}
\description{
Internal function to compare two directories (source and target) and identify files
that are new or modified in the source directory compared to the target directory.
}
\keyword{internal}
