% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-handling_int.R
\name{.fh_onedrive_abs}
\alias{.fh_onedrive_abs}
\title{Resolves relative Path with OneDrive Root Identifier to absolute path}
\usage{
.fh_onedrive_abs(rel_path)
}
\arguments{
\item{rel_path}{Character. The absolute path provided by the user.}
}
\value{
Character. The modified path with the OneDrive root identifier.
}
\description{
This internal function replaces an relative path with a OneDrive root identifier
(`OneDriveConsumer` or `OneDriveCommercial`)
}
\keyword{internal}
