% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{.gen_fields_in}
\alias{.gen_fields_in}
\title{Generate "Fields In" Using a Generic Template}
\usage{
.gen_fields_in(pipe_line)
}
\arguments{
\item{pipe_line}{A \code{data.table} (or similar) with columns
\code{FLDNM_IN} and \code{FIELDTP} at least.}
}
\value{
A single character string with the formatted \code{FLDNM_IN} fields.
}
\description{
Internal wrapper function around \code{.gen_fields_generic} specifically for
fields that reference \code{FLDNM_IN} and \code{FIELDTP}.
}
\keyword{internal}
