% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_data_ingestion.R
\name{.gen_fields_out}
\alias{.gen_fields_out}
\title{Generate "Fields Out" Using a Generic Template}
\usage{
.gen_fields_out(pipe_line)
}
\arguments{
\item{pipe_line}{A \code{data.table} (or similar) with columns
\code{FLDNM_OUT} and \code{TRNSFRM} at least.}
}
\value{
A single character string with the formatted \code{FLDNM_OUT} fields.
}
\description{
Internal wrapper function around \code{.gen_fields_generic} specifically for
fields that reference \code{FLDNM_OUT} and \code{TRNSFRM}.
}
\keyword{internal}
