% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.generate_color_variables}
\alias{.generate_color_variables}
\title{Generate Variables for Ecotone Brand Colors}
\usage{
.generate_color_variables(env = globalenv())
}
\arguments{
\item{env}{The environment where the variables should be created. Defaults to the global environment.}

\item{colors}{A named vector or list of colors.}
}
\value{
NULL. The function creates variables dynamically.
}
\description{
Dynamically creates variables named "col_<name>" in the specified environment.
}
\details{
authorized colours for the logo.
Castle Green will only be used for the housing of the logo.
The colours for the leaves are Forest Green, Light Green and Yellow.
The ecotone wordmark colour is White.
For the tagline only Jet Black or White can be used.
[](https://coolors.co/3e074a-0f431c-fde8e9-f0f600)
}
\keyword{internal}
