% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.generate_environment_variables}
\alias{.generate_environment_variables}
\title{Generate Variables for DTAP Environments}
\usage{
.generate_environment_variables(env = globalenv())
}
\arguments{
\item{env}{The environment where the variables should be created. Defaults to the global environment.}
}
\value{
NULL. The function creates variables dynamically.
}
\description{
Dynamically creates variables named "env_<name>" based on the given list of DTAP environments.
}
\keyword{internal}
