% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.generate_functional_area_variables}
\alias{.generate_functional_area_variables}
\title{Generate Variables for Functional Areas}
\usage{
.generate_functional_area_variables(env = globalenv())
}
\arguments{
\item{env}{The environment where the variables should be created. Defaults to the global environment.}

\item{areas}{A named vector or list of functional areas.}
}
\value{
NULL. The function creates variables dynamically.
}
\description{
Dynamically creates variables named "fun_<name>" based on the given list of functional areas.
}
\keyword{internal}
