% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.generate_stage_variables}
\alias{.generate_stage_variables}
\title{Generate Variables for Stages}
\usage{
.generate_stage_variables(stages, env = globalenv())
}
\arguments{
\item{env}{The environment where the variables should be created. Defaults to the global environment.}
}
\value{
NULL. The function creates variables dynamically.
}
\description{
Dynamically creates variables named "stg_<name>" based on the given list of stages.
}
\keyword{internal}
