% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_functions.R
\name{.get_duckdb_conn}
\alias{.get_duckdb_conn}
\title{Get (or create) a DuckDB connection}
\usage{
.get_duckdb_conn(dbdir = ":memory:")
}
\arguments{
\item{dbdir}{Character. Location of the DuckDB database file. Use \code{":memory:"}
for an in-memory database.}
}
\value{
A `DBIConnection` object to DuckDB.
}
\description{
Internal function that returns a `DBIConnection` to DuckDB. If no connection
exists in the `.duckdb_env`, a new one is created using \code{\link[DBI]{dbConnect}}.
}
\details{
This internal helper function checks if there is an existing DuckDB connection in
the environment `.duckdb_env$conn`. If none is found, it initializes a new one and
caches it. Subsequent calls will reuse the same connection.
}
\keyword{internal}
