% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion_int.R
\name{.get_file_spec}
\alias{.get_file_spec}
\title{Get File Specification from YAML Configuration (Internal)}
\usage{
.get_file_spec(project_dir)
}
\arguments{
\item{project_dir}{Character. The path to the project directory containing the `.config.yaml` file.}
}
\value{
A list with the following elements:
\describe{
  \item{DELIM}{Character. The delimiter used in the files (e.g., `";"`).}
  \item{HEADER}{Logical. Whether the files have a header (`TRUE` or `FALSE`).}
  \item{DATE_FORMAT}{Character. The date format used in the files (e.g., `\"\%Y-\%m-\%d\"`).}
}
}
\description{
Internal function to retrieve the file specification from the YAML configuration file
stored in the specified project directory.
}
\keyword{internal}
