% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_functions.R
\name{.get_parquet_paths}
\alias{.get_parquet_paths}
\title{Return Parquet Paths by VTYPE and FTYPE}
\usage{
.get_parquet_paths(.vtype = c("010", "060"), .ftype = c(1, 2))
}
\arguments{
\item{.vtype}{A character vector of valid vtype codes. Defaults to \code{c("010", "060")}.}

\item{.ftype}{A numeric (or integer) vector of valid ftype codes. Defaults to \code{c(1,2)}.}
}
\value{
A character vector of parquet paths corresponding to all
  \code{(.vtype, .ftype)} pairs in the lookup.
}
\description{
An internal helper function that uses a data.table lookup for valid vtype-ftype-path
combinations. By default, it returns all available paths if no arguments are supplied.
}
\keyword{internal}
