% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_int.R
\name{.hl_config_get}
\alias{.hl_config_get}
\title{Internal function to retrieve most recent config file}
\usage{
.hl_config_get(.reload = FALSE)
}
\arguments{
\item{.reload}{Logical, whether to force reload the config file. Default is FALSE.}
}
\value{
The most recent configuration as a list.
}
\description{
This function checks if the configuration file should be reloaded based on
the file modification time or forced by the `.reload` flag.
}
\keyword{internal}
