% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_int.R
\name{.hl_config_reload}
\alias{.hl_config_reload}
\title{Internal function to retrieve or reload the config}
\usage{
.hl_config_reload(.reload = FALSE)
}
\arguments{
\item{.reload}{Logical, whether to force reload the config file. Default is FALSE.}
}
\value{
The current or reloaded configuration as a list.
}
\description{
This function checks if the configuration file should be reloaded based on
the `.reload` flag or if the file modification time is newer than the cached time.
}
