% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_int.R
\name{.initialize_constants}
\alias{.initialize_constants}
\title{initialize global constants}
\usage{
.initialize_constants()
}
\value{
NULL. This function creates constants dynamically in the global environment.
}
\description{
Creates global constants for colors, functional areas, stages, and environments by calling internal functions.
}
\keyword{internal}
