% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_int.R
\name{.onedrive_reference}
\alias{.onedrive_reference}
\title{Replace Absolute Path with OneDrive Root Identifier}
\usage{
.onedrive_reference(abs_path)
}
\arguments{
\item{abs_path}{Character. The absolute path provided by the user.}
}
\value{
Character. The modified path with the OneDrive root identifier or the original path.
}
\description{
This internal function replaces an absolute path with a OneDrive root identifier
(`OneDriveConsumer` or `OneDriveCommercial`) if the path belongs to the respective
OneDrive directory. If the path does not belong to either directory, the original
path is returned unchanged.
}
\keyword{internal}
