% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{.pa_parquet_describe}
\alias{.pa_parquet_describe}
\title{Describe the Structure of a Parquet File}
\usage{
.pa_parquet_describe(.fn = NULL)
}
\arguments{
\item{.fn}{Character. The file path to the Parquet file.}
}
\value{
Prints the metadata information of the Parquet file to the console.
}
\description{
Retrieves metadata information about a Parquet file by describing its structure using DuckDB.
}
\details{
This function uses DuckDB to describe the structure of a Parquet file. The results are printed to the console.
}
