% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_int.R
\name{.pi_onedrive_get}
\alias{.pi_onedrive_get}
\title{Resolves relative Path with OneDrive Root Identifier to absolute path}
\usage{
.pi_onedrive_get(rel_path)
}
\arguments{
\item{rel_path}{Character. The absolute path provided by the user.}
}
\value{
Character. The modified path with the OneDrive root identifier.
}
\description{
This internal function replaces an relative path with a OneDrive root identifier
(`OneDriveConsumer` or `OneDriveCommercial`)
}
\keyword{internal}
