% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.save_config_to_yaml}
\alias{.save_config_to_yaml}
\title{Save a Configuration to YAML}
\usage{
.save_config_to_yaml(config_list, project_dir = ".")
}
\arguments{
\item{config_list}{A named list of configuration items to write to the YAML file.}

\item{project_dir}{A character string specifying the project directory where
the `.config.yaml` file is (or will be) located. Defaults to `"."`.}
}
\description{
Internal function to save configuration settings to a `.config.yaml` file.
}
\details{
This function is used internally by the package to store key-value pairs
into a YAML file for persistent configuration management.
}
\keyword{internal}
