% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_int.R
\name{.su_config_default_write}
\alias{.su_config_default_write}
\title{Create Default Config YAML File (Internal)}
\usage{
.su_config_default_write(
  config_dir = "./inst/extdata/config",
  key_value_list = NULL
)
}
\arguments{
\item{config_dir}{Character. The directory where the `config.yaml` file will be created. Defaults to `"./inst/extdata"`.}

\item{key_value_list}{List. A named list of key-value pairs to set in the configuration file.}
}
\description{
Internal helper function to create a default `config.yaml` file in the specified directory.
}
\keyword{internal}
