% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_int.R
\name{.su_padt_initialize}
\alias{.su_padt_initialize}
\title{Initialize the padt environment}
\usage{
.su_padt_initialize()
}
\value{
NULL. This function is used for side effects.
}
\description{
This function initializes the `.padt_env` environment for development or testing purposes.
It mimics the `.onLoad` behavior to set up the `config.yaml` path, root directory, and any
other necessary package-specific settings. Use this function during development without needing
to reload the entire package.
}
\details{
- For testing or `devtools::check()` contexts, it sets the root directory to a temporary folder.
- For development within the package directory, it uses `inst/extdata` as the root directory.
- For regular use, it copies the configuration folder to the user's project directory (`"."`).
}
\keyword{internal}
