% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestion.R
\name{.transform_csv_to_parquet}
\alias{.transform_csv_to_parquet}
\title{Transform a Single CSV File to Parquet (Internal)}
\usage{
.transform_csv_to_parquet(
  full_file_name,
  output_path,
  file_spec,
  pipe_line,
  verbose
)
}
\arguments{
\item{full_file_name}{A character string with the full path to the CSV file.}

\item{output_path}{A character string with the directory path where the
Parquet file should be written.}

\item{file_spec}{A list-like object with the CSV specifications (e.g.
delim, header, date format).}

\item{pipe_line}{A \code{data.table} (or similar) detailing how
fields should be transformed.}

\item{verbose}{A logical indicating whether to print additional
information (\code{TRUE}) or not (\code{FALSE}).}
}
\value{
Returns \code{NULL} invisibly. The side effect is creation of a Parquet
file in \code{output_path}.
}
\description{
Internal function that reads a single CSV file, applies transformations,
and writes the result to a Parquet file.
}
\details{
This function:
\enumerate{
  \item Ensures the file is indeed a CSV.
  \item Reads the CSV using \code{duckdb}, applies transformations.
  \item Writes the results to a Parquet file.
}
}
\keyword{internal}
