% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_int.R
\name{.upsert_config_in_yaml}
\alias{.upsert_config_in_yaml}
\title{Upsert (Update/Insert) a Key-Value Pair in Configuration}
\usage{
.upsert_config_in_yaml(.key, .value, .project_dir = ".")
}
\arguments{
\item{.key}{A character string specifying the key to be updated or inserted.}

\item{.value}{The value to assign to the key.}

\item{.project_dir}{A character string specifying the project directory where
the `.config.yaml` file is (or will be) located. Defaults to `"."`.}
}
\description{
Internal function to add or update a single key-value pair in `.config.yaml`.
}
\details{
If the `.config.yaml` file already exists, this function updates or inserts
the key-value pair in that file. If the file does not exist, it is created
along with a minimal configuration list.
}
\keyword{internal}
