% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_int.R
\name{.verbose}
\alias{.verbose}
\title{Display Verbose Output for Function Execution}
\usage{
.verbose(
  function_name = NULL,
  function_args = NULL,
  function_return = NULL,
  verbose = FALSE
)
}
\arguments{
\item{function_name}{Character. The name of the function being executed.}

\item{function_args}{List. The arguments passed to the function. Defaults to \code{NULL}.}

\item{function_return}{Any. The return value of the function. Defaults to \code{NULL}.}

\item{verbose}{Logical. Whether to print the verbose output. Defaults to \code{FALSE}.}
}
\value{
This function does not return anything but prints details to the console if \code{verbose = TRUE}.
}
\description{
Helper function to print details about a function's name, arguments, and return value if verbose mode is enabled.
}
\examples{
# Example for internal testing
}
\keyword{internal}
