% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\docType{data}
\name{DTAP}
\alias{DTAP}
\title{Vector of DTAP Environments}
\format{
An object of class \code{character} of length 4.
}
\usage{
DTAP
}
\description{
These are the default DTAP environments used in the package.
}
\keyword{datasets}
