% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ext.R
\docType{data}
\name{pa_AREA}
\alias{pa_AREA}
\title{Functional pa_AREAs}
\format{
An object of class \code{character} of length 4.
}
\usage{
pa_AREA
}
\description{
Common functional pa_AREAs for demonstration purposes.
}
\keyword{datasets}
