% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-ingestion_ext.R
\name{pa_ask_user_transform_files}
\alias{pa_ask_user_transform_files}
\title{Prompt User for Transformation Decision with Options}
\usage{
pa_ask_user_transform_files(pa_SYNC)
}
\arguments{
\item{pa_SYNC}{A list with two elements:
\describe{
  \item{NEW}{Character vector of paths for new files.}
  \item{MOD}{Character vector of paths for modified files.}
}}
}
\value{
Character. One of the options: ".all", ".new", ".mod", or ".none".
}
\description{
Asks the user how to handle the files found (transform all, transform new, transform modified, or do nothing).
}
