% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plotting_ext.R
\name{pa_brand_color_scale}
\alias{pa_brand_color_scale}
\title{Custom scale_color_manual using brand colors from YAML}
\usage{
pa_brand_color_scale(labels)
}
\arguments{
\item{labels}{Character vector of labels to be used in the legend.}
}
\value{
A ggplot2 scale_color_manual object.
}
\description{
This function retrieves colors from the brand color palette in the YAML configuration
and applies them to ggplot2's scale_color_manual() based on the number of labels.
}
\examples{
\dontrun{
ggplot(data, aes(x, y, color = group)) +
  geom_line() +
  scale_color_brand(labels = c("M-1", "M-2", "M-3"))
  }
}
