% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plotting_ext.R
\name{pa_brand_theme}
\alias{pa_brand_theme}
\title{A ggplot2 theme that applies brand styling from the YAML config.}
\usage{
pa_brand_theme()
}
\value{
A ggplot2 theme object.
}
\description{
A ggplot2 theme that applies brand styling from the YAML config.
}
