% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ext.R
\docType{data}
\name{pa_BSGP}
\alias{pa_BSGP}
\title{pa_BSGP Stages}
\format{
An object of class \code{character} of length 4.
}
\usage{
pa_BSGP
}
\description{
The four data stages commonly used in data processing.
}
\details{
Platinum is used for Forecast Exports
}
\keyword{datasets}
