% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plotting_ext.R
\name{pa_color_brand_scale}
\alias{pa_color_brand_scale}
\title{A ggplot2 theme that applies brand styling from the YAML config.
This theme is suitable for plots with a white background.}
\usage{
pa_color_brand_scale(...)
}
\arguments{
\item{...}{Additional arguments to pass to ggplot2::theme().}
}
\value{
A ggplot2 theme object.
}
\description{
A ggplot2 theme that applies brand styling from the YAML config.
This theme is suitable for plots with a white background.
}
