% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plotting_ext.R
\name{pa_color_get}
\alias{pa_color_get}
\title{Get a color by name or index.}
\usage{
pa_color_get(.name_or_index)
}
\arguments{
\item{.name_or_index}{Either a string (name of the color) or a numeric index.}
}
\value{
A string with the color hex code.
}
\description{
Get a color by name or index.
}
