% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_ext.R
\name{pa_config_set_value}
\alias{pa_config_set_value}
\title{set config value in YAML file}
\usage{
pa_config_set_value(key, value, config_file = ".config.yaml")
}
\arguments{
\item{key}{Character. The key to add or update in the YAML file.}

\item{value}{Any. The value to associate with the given key.}

\item{config_file}{Character. The path to the YAML configuration file. Defaults to ".config.yaml".}
}
\value{
NULL. The function updates the YAML file in place.
}
\description{
Adds or updates a key-value pair in a YAML configuration file.
}
