% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ext.R
\docType{data}
\name{pa_DTAP}
\alias{pa_DTAP}
\title{pa_DTAP Environments}
\format{
An object of class \code{character} of length 5.
}
\usage{
pa_DTAP
}
\description{
These are the default pa_DTAP environments used in the package.
}
\details{
Demo environment is to showcase the package functionality
}
\keyword{datasets}
