% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ext.R
\docType{data}
\name{pa_ET_COLS}
\alias{pa_ET_COLS}
\title{Ecotone Brand Colors}
\format{
An object of class \code{character} of length 6.
}
\usage{
pa_ET_COLS
}
\description{
Ecotone Brand Colors
}
\details{
authorized colours for the logo.
Castle Green will only be used for the housing of the logo.
The colours for the leaves are Forest Green, Light Green and Yellow.
The ecotone wordmark colour is White.
For the tagline only Jet Black or White can be used.
[](https://coolors.co/3e074a-0f431c-fde8e9-f0f600)
}
\keyword{datasets}
