% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{pa_folder_open}
\alias{pa_folder_open}
\title{Open a Folder in the File Explorer}
\usage{
pa_folder_open(path)
}
\arguments{
\item{path}{Character. The path to the folder to open.}
}
\value{
This function does not return anything. It opens the folder in the system's file explorer.
}
\description{
Opens a specified folder in the system's default file explorer (e.g., Windows Explorer on Windows).
}
