% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-access_ext.R
\name{pa_Get_RTP}
\alias{pa_Get_RTP}
\title{Fetch RTP Data from DuckDB}
\usage{
pa_Get_RTP(
  .material = NULL,
  .salesorg = NULL,
  .scope_matl = TRUE,
  .scope_sorg = TRUE,
  .cm_min = "202101",
  .cm_max = "202506",
  .n = Inf
)
}
\arguments{
\item{.material}{[character] (default: `NULL`)
Optional user-supplied material filter. If not specified, all materials are included.}

\item{.salesorg}{[character] (default: `NULL`)
Optional user-supplied sales organization filter. If not specified, all sales organizations are included.}

\item{.scope_matl}{[logical] (default: `TRUE`)
If `TRUE`, restricts data to materials within the Pythia scope.}

\item{.scope_sorg}{[logical] (default: `TRUE`)
If `TRUE`, restricts data to sales organizations within the Pythia scope.}

\item{.cm_min}{[character] (default: `'202101'`)
Specifies the earliest calendar month for which data is available, formatted as `YYYYMM`.}

\item{.cm_max}{[character] (default: `'202506'`)
Specifies the latest calendar month for which data is available, formatted as `YYYYMM`.}

\item{.n}{[numeric] (default: `Inf`)
The maximum number of materials to return. Use `Inf` to return all matching records.}
}
\value{
A `data.table` object containing RTP data with the following columns:
  - `SALESORG`: Sales organization code.
  - `PLANT`: Plant code.
  - `MATERIAL`: Material code.
  - `STEP`: Processing step (fixed as `-1`).
  - `CALMONTH`: Calendar month in `YYYYMM` format.
  - `FTYPE`: Fixed as `'4'`.
  - `VTYPE`: Fixed as `'010'`.
  - `Q`: Sum of sales and free-of-charge quantities.
  - `SLS`: Sum of sales quantities.
  - `RET`: Sum of return quantities.
  - `FOC`: Sum of free-of-charge quantities.
  - `DIR`: Sum of direct sales quantities.
  - `ICS`: Sum of intercompany sales quantities.
  - `MSL`: Sum of MSL quantities.
}
\description{
This function retrieves RTP (Real-Time Planning) data based on specified filters
such as material, sales organization, and calendar month range. It fetches data
from a DuckDB connection and returns it as a `data.table` object.
}
\examples{
# Fetch RTP data for all materials and sales organizations in the Pythia scope
}
