% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{pa_head_parquet}
\alias{pa_head_parquet}
\title{Preview the Header of a Parquet File}
\usage{
pa_head_parquet(.fn, .n = 1000)
}
\arguments{
\item{.fn}{Character. The file path to the Parquet file.}

\item{.n}{Numeric. The number of rows to retrieve. Defaults to 1000.}
}
\value{
A \code{data.table} containing the first \code{n} rows of the Parquet file.
}
\description{
Retrieves the first \code{n} rows of a Parquet file as a \code{data.table}.
}
\details{
This function connects to DuckDB to query the specified Parquet file and fetch the first \code{n} rows.
}
