% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ext.R
\name{pa_initialize}
\alias{pa_initialize}
\title{Run Installation Process for the Package}
\usage{
pa_initialize()
}
\value{
NULL. Functions are executed based on user input.
}
\description{
Guides the user through the package installation process by asking three questions
and executing specific functions (`fun1`, `fun2`, `fun3`) based on the user's choices.
}
\examples{
# Run the installation process
run_installation()
}
