% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants_ext.R
\name{pa_initialize_constants}
\alias{pa_initialize_constants}
\title{Initialize Global Constants}
\usage{
pa_initialize_constants()
}
\value{
NULL. This function creates constants dynamically in the global environment.
}
\description{
Creates global constants for colors, functional areas, stages, and environments by calling internal functions.
}
