% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{pa_MATN1_O}
\alias{pa_MATN1_O}
\title{Remove Leading Zeros}
\usage{
pa_MATN1_O(x)
}
\arguments{
\item{x}{A character vector. Each element will be checked to see if it is purely numeric.}
}
\value{
A character vector with leading zeros removed for numeric strings.
        Non-numeric strings will be returned unchanged.
}
\description{
This function removes leading zeros from a string, but only if the string is purely numeric.
If the string contains non-numeric characters, it will be returned unchanged.
}
\details{
This function mimics the behavior of SAP's \code{CONVERSION_EXIT_MATN1_OUTPUT}
which removes leading zeros from material numbers if these only contain numbers.
}
\examples{
pa_MATN1_O("00123")     # Returns "123"
pa_MATN1_O("00000A123") # Returns "00000A123"
pa_MATN1_O(c("00123", "00000A123", "A000045")) # Returns c("123", "abc123", "A000045")

}
