% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_plotting_ext.R
\name{pa_palette_get}
\alias{pa_palette_get}
\title{Get a palette color from the config by index.}
\usage{
pa_palette_get(.index)
}
\arguments{
\item{.index}{Integer index, e.g. 1, 2, 3...}
}
\value{
A string with the color hex code.
}
\description{
Get a palette color from the config by index.
}
