% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{pa_RL0}
\alias{pa_RL0}
\title{Remove Leading Zeros}
\usage{
pa_RL0(x)
}
\arguments{
\item{x}{A character vector. Each element will be checked to see if it is purely numeric.}
}
\value{
A character vector with leading zeros removed for numeric strings.
        Non-numeric strings will be returned unchanged.
}
\description{
This function removes leading zeros from a string, but only if the string is purely numeric.
If the string contains non-numeric characters, it will be returned unchanged.
}
\examples{
pa_RL0("00123") # Returns "123"
pa_RL0("abc123") # Returns "abc123"
pa_RL0(c("00123", "abc123", "000045")) # Returns c("123", "abc123", "45")

}
