% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ext.R
\name{pa_su_config_reset}
\alias{pa_su_config_reset}
\title{Run installation process for the `padt`-package}
\usage{
pa_su_config_reset()
}
\value{
Message indicating the installation process is completed.
}
\description{
Guides the user through the package installation process.
user can choose to create the 'config' folder, copy the demo environment
and activate the demo environment.
}
\examples{
\dontrun{
# Run the installation process
pa_su_config_reset()
}
}
