% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_ext.R
\name{RL0}
\alias{RL0}
\title{Remove Leading Zeros}
\usage{
RL0(x)
}
\arguments{
\item{x}{A character vector. Each element will be checked to see if it is purely numeric.}
}
\value{
A character vector with leading zeros removed for numeric strings.
        Non-numeric strings will be returned unchanged.
}
\description{
This function removes leading zeros from a string, but only if the string is purely numeric.
If the string contains non-numeric characters, it will be returned unchanged.
}
\examples{
RL0("00123") # Returns "123"
RL0("abc123") # Returns "abc123"
RL0(c("00123", "abc123", "000045")) # Returns c("123", "abc123", "45")

}
