% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pa_functions.R
\name{.get_cte_scope_materials}
\alias{.get_cte_scope_materials}
\title{Construct a Scope Materials CTE Clause}
\usage{
.get_cte_scope_materials(.scope_matl, .con)
}
\arguments{
\item{.con}{A database connection object.}

\item{.apply_scope}{Logical. If \code{TRUE}, returns the scope materials snippet;
otherwise, returns an empty string.}
}
\value{
A character string containing the SQL snippet for scope materials.
}
\description{
Internal helper function that returns a SQL snippet (CTE) for scope materials 
if \code{.apply_scope} is \code{TRUE}, otherwise returns an empty string.
}
\keyword{internal}
