% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_int.R
\name{.su_data_folders_create}
\alias{.su_data_folders_create}
\title{Create Data Folder Structure (Internal)}
\usage{
.su_data_folders_create(root_dir, .bsgp = .BSGP, .area = .AREA, .dtap = .DTAP)
}
\arguments{
\item{root_dir}{Character. The root directory where the folder structure will be created.}

\item{.bsgp}{Character vector. The Bronze, Silver, Gold, Platinum tiers, defaults to `.BGSP`.}

\item{.area}{Character vector. The functional areas, defaults to `.AREA`.}

\item{.dtap}{Character vector. The DTAP stages, defaults to `.DTAP`.}
}
\value{
NULL. The function creates the directory structure and does not return anything.
}
\description{
Internal helper function to create a folder structure based on the root directory,
BS(GP) tiers, functional areas, and DTAP stages.
}
\keyword{internal}
