% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml_ext.R
\name{pa_config_get_value}
\alias{pa_config_get_value}
\title{get config value from YAML file}
\usage{
pa_config_get_value(key, config_file = ".config.yaml")
}
\arguments{
\item{key}{Character. The key whose value needs to be retrieved.}

\item{config_file}{Character. The path to the YAML configuration file. Defaults to ".config.yaml".}
}
\value{
The value associated with the provided key, or NULL if the key does not exist.
}
\description{
Reads a value from a YAML configuration file based on the provided key.
}
