% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths_ext.R
\name{pa_getwd}
\alias{pa_getwd}
\title{Get Working Directory}
\usage{
pa_getwd(project_dir)
}
\arguments{
\item{project_dir}{Character. The path to the project directory containing the `.config.yaml` file.}
}
\value{
Character. The absolute path to the working directory.
}
\description{
Determines the working directory based on the project directory and the configuration file.
resolves placeholders such as `"OneDriveConsumer"` or `"OneDriveBusiness"`
back to real file paths, then appends the currently set environment to the path.
}
