% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_ext.R
\name{pa_setup_project_structure}
\alias{pa_setup_project_structure}
\title{Set Up a Project Directory Structure}
\usage{
pa_setup_project_structure(
  root_dir = ".",
  pa_DTAP_levels = pa_DTAP,
  pa_BSGP_levels = pa_BSGP,
  functional_pa_AREAs = pa_AREA,
  project_dir = "."
)
}
\arguments{
\item{root_dir}{A character string specifying the top-level directory under
which all subdirectories will be created. Defaults to `"."`.}

\item{pa_DTAP_levels}{A character vector defining the pa_DTAP levels.
Defaults to `pa_DTAP` = c("Development", "Test", "Acceptance", "Production").}

\item{pa_BSGP_levels}{A character vector defining the pa_BSGP stages.
Defaults to `pa_BSGP` = c("Bronze", "Silver", "Gold", "Platinum").}

\item{functional_pa_AREAs}{A character vector defining functional pa_AREAs within
each pa_BSGP stage. Defaults to
`pa_AREA` = c("sales", "stock", "promotions", "master_data").}

\item{project_dir}{A character string specifying the directory where the
`.config.yaml` file will be saved or updated. Defaults to `"."`.}
}
\value{
Invisibly returns the normalized `root_dir`.
}
\description{
Creates a directory structure for different pa_DTAP environments, pa_BSGP stages,
and functional pa_AREAs, and saves the root directory location in `.config.yaml`.
}
\details{
This function constructs a multi-level directory hierarchy reflecting
different pa_DTAP (Development, Test, Acceptance, Production) environments,
each with pa_BSGP (Bronze, Silver, Gold, Platinum) stages and designated
functional pa_AREAs (e.g., "sales", "stock"). If the specified `root_dir` is
within the user's OneDrive path (either consumer or commercial), the path
stored in the YAML file will be replaced with a placeholder
("OneDriveConsumer" or "OneDriveBusiness") and the relative subdirectory
path.
}
\seealso{
[fs::dir_create()] for directory creation,
  [fs::path_abs()] and [fs::path_rel()] for path manipulations,
  [yaml::write_yaml()] for saving configurations.
}
